Split out legacy TRN journey tests
Also un-obsolete TRN scope; having to ignore the obsoletion warnings
everywhere is painful.
