+++ dotnet-authserver/src/TeacherIdentity.AuthServer/Oidc/CustomScopes.cs
@@ -6,7 +6,6 @@ public static class CustomScopes
     public const string UserRead = "user:read";
     public const string UserWrite = "user:write";
     public const string DqtRead = "dqt:read";
-    [Obsolete("Use DqtRead instead.")]
     public const string Trn = "trn";
 
     public static string[] All => StaffUserTypeScopes.Concat(DefaultUserTypesScopes).ToArray();
@@ -21,8 +20,6 @@ public static class CustomScopes
     public static string[] DefaultUserTypesScopes { get; } = new[]
     {
         DqtRead,
-#pragma warning disable CS0618 // Type or member is obsolete
         Trn
-#pragma warning restore CS0618 // Type or member is obsolete
     };
 }
