+++ dotnet-authserver/tests/TeacherIdentity.AuthServer.EndToEndTests/appsettings.json
@@ -36,7 +36,8 @@
       "PostLogoutRedirectUris": [
         "http://localhost:55342/oidc/signout-callback"
       ],
-      "Scopes": [ "trn", "user:read", "user:write", "dqt:read" ]
+      "Scopes": [ "trn", "user:read", "user:write", "dqt:read" ],
+      "TrnRequirementType": "Required"
     },
     {
       "ClientId": "testcc",
